update "redirect" hook function: support Object return value
previously, it was assumed that the return value is a string;
the value of this string is a new URL

now, the return value can be a string.. same as before;
but, it can also be an Object: {matching_url, file_name, file_ext}

attributes are optional; the value will only be updated when present.
in most cases, returning an Object will not be necessary.
however, if the purpose of the redirect is to trick the caching
mechanism into prefetching URLs that don't have a ".ts" extension,
then the return value would need to be an Object such that:
* "matching_url" contains ".ts"
  * this can be achieved by appending a hash to the URL
    * ex: "#chunk.ts"
* "file_ext" === ".ts"